requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.7.2
    # via django
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
django==3.2.25
    # via
    #   -r requirements.in
    #   django-filter
    #   django-jsonfield
    #   djangorestframework
    #   drf-spectacular
django-filter==23.5
    # via -r requirements.in
django-jsonfield==1.4.1
    # via -r requirements.in
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   drf-spectacular
drf-spectacular==0.27.1
    # via -r requirements.in
importlib-metadata==7.0.2
    # via markdown
importlib-resources==6.3.0
    # via
    #   jsonschema
    #   jsonschema-specifications
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.21.1
    # via drf-spectacular
jsonschema-specifications==2023.12.1
    # via jsonschema
markdown==3.6
    # via -r requirements.in
mysqlclient==2.1.1
    # via -r requirements.in
pkgutil-resolve-name==1.3.10
    # via jsonschema
pytz==2024.1
    # via django
pyyaml==6.0.1
    # via drf-spectacular
referencing==0.33.0
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.18.0
    # via
    #   jsonschema
    #   referencing
six==1.16.0
    # via django-jsonfield
sqlparse==0.4.4
    # via django
typing-extensions==4.10.0
    # via
    #   asgiref
    #   drf-spectacular
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.18
    # via -r requirements.in
zipp==3.18.1
    # via
    #   importlib-metadata
    #   importlib-resources
